Modularize fabric project
