+++ src/fabric_project/config.py
@@ -1,0 +1,3 @@
+NUM_CLIENTS = 4
+EPOCHS = 2
+BATCH_SIZE = 16
